+++ lib/aws_csshx.rb
@@ -1,2 +1,7 @@
+require 'etc'
+require 'right_aws'
+
 require 'aws_csshx/application'
 require 'aws_csshx/configuration_file'
+require 'aws_csshx/misc'
+require 'aws_csshx/options'