+++ Changelog.md
@@ -1,3 +1,8 @@
-### colortail 0.1.0 2012-05-01
+### aws_csshx 0.1.1 2012-05-04
+
+ * Added command line switches (Eric Lubow)
+ * Added config file support using .csshrc (Eric Lubow)
+
+### aws_csshx 0.1.0 2012-05-01
 
  * Initial vesion (Russell Bradberry)